chore: Rename `new` command to `nfb`
