+++ src/commands/mod.rs
@@ -1,2 +1,2 @@
-pub mod new;
+pub mod nfb;
 pub mod finish;
